Add pressure metric for bpf maps
Adds a metric to tell the fill percentage from 0 to 1 of maps filled
from userspace. This metric is a prometheus const label gauge that is
registered and unregistered dynamically for each map when over/under
a threshold respectively. By default the threshold is 0.0 for all maps
except the policy map which is set to 0.1.